agents/broski-bot/pyproject.toml
[tool.poetry]
name = "broski-bot"
version = "4.0.0"
description = "Production-grade neurodivergent-friendly Discord automation bot"
authors = ["Lyndz Williams <lyndz@example.com>"]
license = "MIT"
readme = "README.md"
homepage = "https://github.com/welshDog/BROski-Bot"
repository = "https://github.com/welshDog/BROski-Bot"
keywords = ["discord", "bot", "productivity", "adhd", "neurodivergent"]

[tool.poetry.dependencies]
python = "^3.11"
discord-py = "^2.3.2"
pydantic = "^2.5.0"
pydantic-settings = "^2.1.0"
sqlalchemy = {extras = ["asyncio"], version = "^2.0.23"}
asyncpg = "^0.29.0"
alembic = "^1.13.0"
redis = {extras = ["hiredis"], version = "^5.0.1"}
aiohttp = "^3.9.1"
python-dotenv = "^1.0.0"
structlog = "^24.1.0"
prometheus-client = "^0.19.0"
sentry-sdk = "^1.39.2"
python-json-logger = "^2.0.7"
tenacity = "^8.2.3"
click = "^8.1.7"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
pytest-asyncio = "^0.21.1"
pytest-cov = "^4.1.0"
pytest-mock = "^3.12.0"
pytest-xdist = "^3.5.0"
faker = "^22.0.0"
black = "^23.12.1"
isort = "^5.13.2"
flake8 = "^7.0.0"
mypy = "^1.7.1"
bandit = "^1.7.6"
pre-commit = "^3.6.0"
pylint = "^3.0.3"
coverage = {extras = ["toml"], version = "^7.4.0"}
safety = "^2.3.5"

[tool.poetry.scripts]
broski = "src.main:cli"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ['py311']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
line_length = 100
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
strict_equality = true
strict_concatenate = true

[[tool.mypy.overrides]]
module = "discord.*"
ignore_missing_imports = true

[tool.pytest.ini_options]
minversion = "7.0"
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "-ra",
    "--strict-markers",
    "--strict-config",
    "--cov=src",
    "--cov-branch",
    "--cov-report=term-missing:skip-covered",
    "--cov-report=html",
    "--cov-report=xml",
    "--cov-fail-under=80",
]

[tool.coverage.run]
source = ["src"]
branch = true
omit = [
    "*/tests/*",
    "*/migrations/*",
    "*/__pycache__/*",
]

[tool.coverage.report]
precision = 2
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "@abstractmethod",
]

[tool.bandit]
exclude_dirs = ["/tests", "/migrations"]
tests = ["B201", "B301"]
skips = ["B101", "B601"]
